+++ lib/src/ui/board/board_builder.dart
@@ -1,7 +1,7 @@
 part of 'board.dart';
 
 /// A function to build a widget from [rank], [file] and [squareSize].
-typedef SquareBuilder = Widget Function(int rank, int file, double squareSize);
+typedef SquareBuilder = Widget? Function(int rank, int file, double squareSize);
 
 ///  Builds a grid of widgets of [size], according to [builder].
 class BoardBuilder extends StatelessWidget {
@@ -11,10 +11,14 @@ class BoardBuilder extends StatelessWidget {
   /// The size of the board.
   final BoardSize size;
 
+  /// If false, elements will not necessarily be aligned to their squares.
+  final bool forceSquareAlignment;
+
   const BoardBuilder({
     super.key,
     required this.builder,
     this.size = BoardSize.standard,
+    this.forceSquareAlignment = true,
   });
 
   @override
@@ -27,12 +31,16 @@ class BoardBuilder extends StatelessWidget {
           return Column(
             children: List.generate(
               size.v,
-              (rank) => Expanded(
-                child: Row(
-                  children: List.generate(
-                    size.h,
-                    (file) => builder(rank, file, squareSize),
-                  ),
+              (rank) => Row(
+                children: List.generate(
+                  size.h,
+                  (file) => forceSquareAlignment
+                      ? SizedBox(
+                          width: squareSize,
+                          height: squareSize,
+                          child: builder(rank, file, squareSize),
+                        )
+                      : builder(rank, file, squareSize) ?? Container(),
                 ),
               ),
             ),